feat: refactor to use new abstract config
