+++ src/services/server_manager.py
@@ -1,12 +1,11 @@
 from abc import ABC, abstractmethod
 
-from schemas.minecraft_server_itzg_config import MinecraftServerITZGConfig
-
+from schemas.minecraft_server_config.minecraft_server_config import MinecraftServerConfig
 
 class ServerManager(ABC):
 
     @abstractmethod
-    def create(self, config: MinecraftServerITZGConfig) -> str | None:
+    def create(self, config: MinecraftServerConfig) -> str | None:
         """Creates a server and return the unique identifier for it"""
         pass
